This pc side and client side codes serve the SPAKE math in integer
I used int to get rid of BigInteger and this code contains lot of things that is unnecessary or not yet implemented. The functionality served is just of SPAKE mathematics using ecdh. Its basic and may be flop sometimes :P :P
The length of the parameters can be increased later.. :p